// lib/api/model/modelsUser.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BVR STR User API
 * API for BVR STR user management and rewards system
 * OpenAPI spec version: 1.0
 */
import type { ModelsAddress } from './modelsAddress';
import type { ModelsOrder } from './modelsOrder';
import type { ModelsReward } from './modelsReward';

export interface ModelsUser {
  address: ModelsAddress;
  createdAt?: string;
  deletedAt?: string;
  email: string;
  id?: string;
  name: string;
  orders?: ModelsOrder[];
  phone: string;
  points?: number;
  rewards?: ModelsReward[];
  updatedAt?: string;
}